Move alignment commands and helpers into own file
